requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
xgboost>=1.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
pyyaml>=6.0
shap>=0.43.0
joblib>=1.3.0
tqdm>=4.66.0

